+++ entry_types/paged/app/controllers/pageflow_paged/editor/entries_controller.rb
@@ -2,6 +2,10 @@ module PageflowPaged
   module Editor
     # @api private
     class EntriesController < ActionController::Base
+      include Pageflow::EditorController
+
+      skip_before_action :verify_edit_lock, only: :partials
+
       helper Pageflow::AudioFilesHelper
       helper Pageflow::BackgroundImageHelper
       helper Pageflow::EntriesHelper
@@ -19,8 +23,14 @@ module PageflowPaged
       helper Pageflow::SocialShareLinksHelper
       helper Pageflow::TextDirectionHelper
       helper Pageflow::VideoFilesHelper
+      helper Pageflow::WidgetsHelper
 
       helper_method :render_to_string
+
+      def partials
+        I18n.locale = @entry.locale
+        render action: 'partials', layout: false
+      end
     end
   end
 end
