+++ entry_types/paged/app/views/pageflow_paged/editor/entries/partials.html.erb
