+++ entry_types/paged/config/routes.rb
@@ -1,0 +1,3 @@
+PageflowPaged::Engine.routes.draw do
+  get '/partials' => 'editor/entries#partials'
+end
