+++ entry_types/paged/packages/pageflow-paged/src/editor/views/EntryPreviewView.js
@@ -144,7 +144,7 @@ export const EntryPreviewView = Marionette.ItemView.extend({
   fetchWidgets: function() {
     var view = this;
 
-    $.ajax('/entries/' + this.model.get('id') + '/partials').success(function(response) {
+    $.ajax(this.model.url() + '/paged/partials').success(function(response) {
       var partials = $('<div />').html(response);
 
       view.ui.header.replaceWith(partials.find('> .header'));
