+++ spec/controllers/pageflow/entries_controller_spec.rb
@@ -136,84 +136,6 @@ module Pageflow
       end
     end
 
-    describe '#partials' do
-      it 'reponds with success for previewers of the entry' do
-        user = create(:user)
-        entry = create(:entry, with_previewer: user)
-
-        sign_in(user, scope: :user)
-        get(:partials, params: {id: entry})
-
-        expect(response.status).to eq(200)
-      end
-
-      it 'requires the signed in user to be previewer of the parent entry' do
-        user = create(:user)
-        entry = create(:entry)
-
-        sign_in(user, scope: :user)
-        get(:partials, params: {id: entry})
-
-        expect(response).to redirect_to(main_app.admin_root_path)
-      end
-
-      it 'requires authentication' do
-        entry = create(:entry)
-
-        get(:partials, params: {id: entry})
-
-        expect(response).to redirect_to(main_app.new_user_session_path)
-      end
-
-      it 'renders editor enabled widgets for entry' do
-        widget_type =
-          TestWidgetType.new(name: 'test_widget',
-                             enabled_in_editor: true,
-                             rendered: '<div class="test_widget"></div>')
-        non_editor_widget_type =
-          TestWidgetType.new(name: 'non_editor_widget',
-                             enabled_in_editor: false,
-                             rendered: '<div class="non_editor_widget"></div>')
-
-        pageflow_configure do |config|
-          config.widget_types.register(widget_type)
-          config.widget_types.register(non_editor_widget_type)
-        end
-
-        user = create(:user)
-        entry = create(:entry, with_previewer: user)
-        create(:widget, subject: entry.draft, role: 'header', type_name: 'test_widget')
-        create(:widget, subject: entry.draft, role: 'footer', type_name: 'non_editor_widget')
-
-        sign_in(user, scope: :user)
-        get(:partials, params: {id: entry})
-
-        expect(response.body).to have_selector('div.test_widget')
-        expect(response.body).not_to have_selector('div.non_editor_widget')
-      end
-
-      it 'uses locale of entry' do
-        widget_type =
-          TestWidgetType.new(name: 'test_widget',
-                             enabled_in_editor: true,
-                             rendered: lambda { %'<div lang="#{I18n.locale}"></div>' })
-
-        pageflow_configure do |config|
-          config.widget_types.register(widget_type)
-        end
-
-        user = create(:user)
-        entry = create(:entry, with_previewer: user)
-        create(:widget, subject: entry.draft, type_name: 'test_widget')
-        entry.draft.update(locale: 'de')
-
-        sign_in(user, scope: :user)
-        get(:partials, params: {id: entry})
-
-        expect(response.body).to have_selector('div[lang=de]')
-      end
-    end
-
     describe '#page' do
       it 'redirects to entry path with page perma id anchor' do
         entry = create(:entry, :published, title: 'report')
